Migrate check-googlesans.py for fontbakery 0.12
